+++ app.js
@@ -2,6 +2,7 @@ import express, { json } from 'express'
 import cors from 'cors'
 import usersRoutes from './routes/users.js'
 import beersRoutes from './routes/beers.js'
+import ratingsRoutes from './routes/ratings.js'
 
 const port = process.env.PORT || 8000
 
@@ -12,6 +13,7 @@ app.use(cors({
     origin: process.env.ALLOWED_ORIGINS.split(',')
 }))
 
+app.use('/users/ratings', ratingsRoutes)
 app.use('/users', usersRoutes)
 app.use('/beers', beersRoutes)
